Add initiative property to allow group by
Add a new `initiative` property to the `TimeEntry` model. I build this
new  property using the text before `:` in the description, or the
full description if there are no `:`.

I tried to create it as a field in the `TimeEntry` model, but it would
require a switch from `BaseModel` to `@dataclass`, and I couldn't use
the `parse_raw()` helper anymore.

This result is fine, I'll re-evaluate this implementation in the
future.

refs: #9