Add the --format option with 'text' and 'ndjson' (#11)
* Add the --format option with 'text' and 'ndjson'

The CLI now offers a new `--format` option that supports the following
formats:

- `text`
- `ndjson`

The `text` format renders the result using tables. This is the default
format.

The `ndjson` format renders the result as JSON objects separated by a
"\n". This is useful for using the data with other tools.

***

The `Result` classes now must support a `ndjson()` method alongside
the current `str()` method. The `ndjson()` method must return an NDJSON
representation of the data.

We added a new `result.render()` function for the `cli` module. This
the function will take care of calling the proper method on the `Result`
object, according to the `format` argument.

refs: #7

* Update README with an example of --format option